Move out of classifier
